Change query in codelists.GetCodeDatasets method to fetch edge info also
The query must not only scrutinize the "usedBy" edges between Code and CodeList
vertices, but it must also harvest the "label" property from that edge, as
well as the "Instance" vertices. So update the query to make this possible.

The returned graphson now has heterogeneous types and cannot be parsed by
gedge.graphson - will run the same query TWICE, first to get the vertices,
and second to get the instances.

Not finished - just a back up of WIP at this stage.

A nastly cludge - to avoid spending more time on trying to get the query to
send back just strings from both sources.